--- src/App.css
@@ -1,14 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  font-family: system-ui;
-  box-sizing: border-box;
-}
-
-#root {
-  background-color: #0d1117;
-  display: grid;
-  place-content: center;
-  min-height: 100vh;
-  font-size: 14px;
-}
