+++ src/components/App/App.css
@@ -1,0 +1,20 @@
+* {
+  margin: 0;
+  padding: 0;
+  font-family: system-ui;
+  box-sizing: border-box;
+}
+
+:root {
+  --theme-background: #0d1117;
+}
+
+body {
+  background-color: var(--theme-background);
+}
+
+body,
+input,
+button {
+  font-size: 14px;
+}
