+++ src/components/Field/Field.css
@@ -20,9 +20,11 @@
   outline: none;
   color: #fff;
   z-index: 1;
+  transition: transform 0.3s;
 }
 
 .Field__container {
+  flex: auto;
   position: relative;
   display: flex;
   flex-direction: column;
